improve/simplify script that takes care of copying the 'results' files over

+++ results/copy_all_results.sh
@@ -1,25 +1,16 @@
 #!/bin/bash -v
 
-# experiment folder that needs to be copied to local machine. All folders (*) if not specified
-EXP=${1:-"*"}
+DIRS="$(ssh abdel@magni.inf.ed.ac.uk 'ls /fs/bil0/abdel/parallel-corpus-filtering/results')"
 
-rsync -au --progress \
-    abdel@magni.inf.ed.ac.uk:/fs/bil0/abdel/parallel-corpus-filtering/results/$EXP .
-
-if [ "$EXP" != "*" ]; then
-    find ./$EXP -name "*.bpe" \
-        -o -name "*.de" \
-        -o -name "*.en" \
-        -o -name "*.json" \
-        -o -name "*.npz" \
-        | tee >(xargs rm -f) >(tar -czf ./$EXP/exp_files.tar.gz -T -)
-else
-    for d in */ ; do
-        find ./$d -name "*.bpe" \
+for dir in $DIRS ; do
+    if ! test -f $dir/exp_files.tar.gz; then
+        rsync -au --progress \
+            abdel@magni.inf.ed.ac.uk:/fs/bil0/abdel/parallel-corpus-filtering/results/$dir .
+        find ./$EXP -name "*.bpe" \
             -o -name "*.de" \
             -o -name "*.en" \
             -o -name "*.json" \
             -o -name "*.npz" \
-            | tar -czf ./$d/exp_files.tar.gz -T -
-    done
-fi
+            | tee >(xargs rm -f) >(tar -czf ./$dir/exp_files.tar.gz -T -)
+    fi
+done
